+++ src/main/java/lotto/LottoMain.java
@@ -1,5 +1,13 @@
 package lotto;
 
+import lotto.domain.game.LottoGameCount;
+import lotto.domain.game.LottoRevenue;
+import lotto.domain.number.LottoNumbers;
+import lotto.domain.number.NumberGenerateStrategy;
+import lotto.domain.number.RandomGenerateStrategy;
+import lotto.domain.ranking.LottoRankingCount;
+import lotto.domain.game.WinnerTicket;
+
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
@@ -23,7 +31,5 @@ public class LottoMain {
         LottoRankingCount lottoRankingCount = LottoRankingCount.of(lottoTickets, winnerTicket);
 
         double lottoRevenueRate = LottoRevenue.from(lottoRankingCount).calculateRevenueRate(lottoGameCount);
-
-
     }
 }
