+++ src/main/java/lotto/domain/game/LottoGameCount.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain.game;
 
 public class LottoGameCount {
