+++ src/main/java/lotto/domain/game/WinnerTicket.java
@@ -1,4 +1,8 @@
-package lotto;
+package lotto.domain.game;
+
+import lotto.domain.number.LottoNumber;
+import lotto.domain.number.LottoNumbers;
+import lotto.domain.number.ManualInputStrategy;
 
 import java.util.List;
