+++ src/main/java/lotto/domain/number/LottoNumbers.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain.number;
 
 import java.util.HashSet;
 import java.util.List;
