+++ src/main/java/lotto/domain/number/ManualInputStrategy.java
@@ -1,8 +1,10 @@
-package lotto;
+package lotto.domain.number;
+
+import lotto.domain.number.NumberGenerateStrategy;
 
 import java.util.List;
 
-public class ManualInputStrategy implements NumberGenerateStrategy{
+public class ManualInputStrategy implements NumberGenerateStrategy {
     private final List<Integer> manualInputNumbers;
 
     public ManualInputStrategy(List<Integer> manualInputNumbers){
