+++ src/main/java/lotto/domain/number/NumberGenerateStrategy.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain.number;
 
 import java.util.List;
