+++ src/test/java/lotto/domain/game/LottoRevenueTest.java
@@ -1,11 +1,12 @@
-package lotto;
+package lotto.domain.game;
 
+import lotto.domain.game.LottoRevenue;
+import lotto.domain.ranking.LottoRanking;
+import lotto.domain.ranking.LottoRankingCount;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import java.util.Arrays;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -16,7 +17,7 @@ public class LottoRevenueTest {
     @Test
     void create() {
         //given
-        Map<LottoRanking , Integer> rankingCount = new HashMap<>();
+        Map<LottoRanking, Integer> rankingCount = new HashMap<>();
         rankingCount.put(LottoRanking.RANK_4,1);
         rankingCount.put(LottoRanking.RANK_5,2);
