+++ src/test/java/lotto/domain/number/LottoNumberTest.java
@@ -1,5 +1,6 @@
-package lotto;
+package lotto.domain.number;
 
+import lotto.domain.number.LottoNumber;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
